Convert ChordsInRealSongsExercise.ts
