+++ src/app/exercise/Exercise.ts
@@ -261,7 +261,7 @@ export namespace Exercise {
     readonly id: string;
     readonly name: string;
     readonly summary: string;
-    readonly explanation: ExerciseExplanationContent;
+    readonly explanation?: ExerciseExplanationContent;
     readonly blackListPlatform?: Platforms;
 
     getAnswerList(): AnswerList<GAnswer>;
